backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
pillow==11.1.0
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
stripe==11.5.0
typing_extensions==4.12.2
urllib3==2.3.0
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cssselect2==0.7.0
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
fonttools==4.56.0
idna==3.10
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
requests==2.32.3
sqlparse==0.5.3
stripe==11.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
urllib3==2.3.0
weasyprint==64.1
webencodings==0.5.1
zopfli==0.2.3.post1
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cssselect2==0.7.0
Django==5.1.6
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-js-asset==3.1.0
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
fonttools==4.56.0
idna==3.10
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
requests==2.32.3
sqlparse==0.5.3
stripe==11.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
urllib3==2.3.0
weasyprint==64.1
webencodings==0.5.1
zopfli==0.2.3.post1
